Ch.2 Programming Project 5
Write a program that determines the change to be dispensed from a vending machine. An item in the machine can cost between 25 cents and a dollar, in 5-cent increments, and the machine accepts only a single dollar bill to pay for the item. For example, a possible dialogue with the user might be
Enter price of item
(from 25 cents to a dollar, in 5-cent increments): 45

You bought an item for 45 cents and gave me a dollar,
so your change is
2 quarters,
0 dimes, and
1 nickle.