Ch. 2 Programming Project 6
Write a program that reads a 4-bit binary number from the keyboard as a string and then converts it to a decimal. For example, if the input is 1100, the output shudl be 12. (Hint: Break the string into substrings and then convert each substring to a value for a single bit. If the bits are b0, b1, b2, and b3, the decimal equivalent is 8b0 + 4b1 + 2b2 + b3.)